requirements.txt
Django==2.2.1
python-dotenv==0.10.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
PyJWT==1.7.1

